Move to altcoin package, add license
