+++ gui/src/main/java/io/bitsquare/gui/util/validation/altcoins/ByteballAddressValidator.java
@@ -1,5 +1,23 @@
-package io.bitsquare.gui.util.validation;
+/*
+ * This file is part of Bitsquare.
+ *
+ * Bitsquare is free software: you can redistribute it and/or modify it
+ * under the terms of the GNU Affero General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or (at
+ * your option) any later version.
+ *
+ * Bitsquare is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+ * FITNESS FOR A PARTICULAR PURPOSE. See the GNU Affero General Public
+ * License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with Bitsquare. If not, see <http://www.gnu.org/licenses/>.
+ */
 
+package io.bitsquare.gui.util.validation.altcoins;
+
+import io.bitsquare.gui.util.validation.InputValidator;
 import org.apache.commons.codec.binary.Base32;
 import org.apache.commons.codec.binary.Base64;
